Rename the app to Friendly Plans.
