+++ .env.production
@@ -1,4 +1,4 @@
-PACKAGE_NAME=com.stxnext.rnstarterkit
+PACKAGE_NAME=com.friendlyplans
 IS_RELEASE=true
 APP_VERSION_CODE=1
 APP_VERSION_NAME=1.0.0
