+++ .env.staging
@@ -1,4 +1,4 @@
-PACKAGE_NAME=com.stxnext.rnstarterkit.staging
+PACKAGE_NAME=com.friendlyplans.staging
 IS_RELEASE=true
 APP_VERSION_CODE=1
 APP_VERSION_NAME=1.0.0
