+++ android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.stxnext.rnstarterkit">
+    package="com.friendlyplans">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.SYSTEM_ALERT_WINDOW" />
@@ -8,7 +8,7 @@
     <uses-permission android:name="android.permission.VIBRATE" />
 
     <application
-        android:name="com.stxnext.rnstarterkit.MainApplication"
+        android:name="com.friendlyplans.MainApplication"
         android:allowBackup="false"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -27,7 +27,7 @@
             android:name="firebase_crashlytics_collection_enabled"
             android:value="false" />
         <activity
-            android:name="com.stxnext.rnstarterkit.MainActivity"
+            android:name="com.friendlyplans.MainActivity"
             android:configChanges="keyboard|keyboardHidden|orientation|screenSize|layoutDirection|locale"
             android:label="@string/app_name"
             android:launchMode="singleTop"
