+++ android/app/src/main/java/com/friendlyplans/MainApplication.java
@@ -1,4 +1,4 @@
-package com.stxnext.rnstarterkit;
+package com.friendlyplans;
 
 import android.app.Application;
