+++ index.js
@@ -3,4 +3,4 @@
 import { AppRegistry } from 'react-native';
 import App from './src/App';
 
-AppRegistry.registerComponent('ReactNativeStarterKit', () => App);
+AppRegistry.registerComponent('FriendlyPlans', () => App);
